moving sync setup to the Application extension class to not have to worry about multiple calls; also set up receiver to recieve connectivity change broadcasts
