+++ openmrs-client/src/main/java/org/openmrs/mobile/activities/ACBaseActivity.java
@@ -68,18 +68,6 @@ public abstract class ACBaseActivity extends AppCompatActivity implements Naviga
 	private Toolbar toolbar;
 	private OpenMRS instance = OpenMRS.getInstance();
 
-	// Sync constants
-	// The authority for the sync adapter's content provider
-	public static final String AUTHORITY = "org.openmrs.mobile.provider";
-	// An account type, in the form of a domain name
-	public static final String ACCOUNT_TYPE = "org.openmrs.mobile.datasync";
-	// The account name (this isn't used by the app, but is needed for syncing)
-	public static final String ACCOUNT = "defaultAccount";
-	// Sync interval constants
-	public static final long SECONDS_PER_MINUTE = 60L;
-	public static final long SYNC_INTERVAL_IN_MINUTES = 1L;
-	public static final long SYNC_INTERVAL = SYNC_INTERVAL_IN_MINUTES * SECONDS_PER_MINUTE;
-
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -89,7 +77,6 @@ public abstract class ACBaseActivity extends AppCompatActivity implements Naviga
 		frameLayout = (FrameLayout)findViewById(R.id.content_frame);
 		intitializeToolbar();
 		intitializeNavigationDrawer();
-		initializeDataSync();
 	}
 
 	@Override
@@ -131,50 +118,6 @@ public abstract class ACBaseActivity extends AppCompatActivity implements Naviga
 		mOpenMRS.clearUserPreferencesData();
 		mAuthorizationManager.moveToLoginActivity();
 		finish();
-	}
-
-	private void initializeDataSync() {
-		Boolean isDataSyncInitialized = ((OpenMRS) this.getApplication()).getIsDataSyncInitialized();
-		if (isDataSyncInitialized) {
-			return;
-		}
-		// Create the dummy account
-		Account mAccount = createSyncAccount(this);
-		// Turn on periodic syncing
-		ContentResolver.setSyncAutomatically(mAccount, AUTHORITY, true);
-		ContentResolver.addPeriodicSync(mAccount, AUTHORITY, Bundle.EMPTY, 3600);
-		ContentResolver.requestSync(mAccount, AUTHORITY, Bundle.EMPTY);
-		((OpenMRS) this.getApplication()).setIsDataSyncInitialized(true);
-	}
-
-	/**
-	 * Create a new dummy account for the sync adapter
-	 *
-	 * @param context The application context
-	 */
-	public static Account createSyncAccount(Context context) {
-		// Create the account type and default account
-		Account newAccount = new Account(ACCOUNT, ACCOUNT_TYPE);
-		// Get an instance of the Android account manager
-		AccountManager accountManager = (AccountManager) context.getSystemService(ACCOUNT_SERVICE);
-
-		// Below needed for syncing, but is for dummy account
-		if (accountManager.addAccountExplicitly(newAccount, null, null)) {
-            /*
-             * If you don't set android:syncable="true" in
-             * in your <provider> element in the manifest,
-             * then call context.setIsSyncable(account, AUTHORITY, 1)
-             * here.
-             */
-			// Intentionally left blank
-		} else {
-            /*
-             * The account exists or some other error occurred. Log this, report it,
-             * or handle it internally.
-             */
-			// Intentionally left blank
-		}
-		return newAccount;
 	}
 
 	private void showLogoutDialog() {
