+++ openmrs-client/src/main/java/org/openmrs/mobile/application/OpenMRS.java
@@ -14,9 +14,18 @@
 
 package org.openmrs.mobile.application;
 
+import android.accounts.Account;
+import android.accounts.AccountManager;
 import android.app.Application;
+import android.content.BroadcastReceiver;
+import android.content.ContentResolver;
+import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.SharedPreferences;
+import android.net.ConnectivityManager;
 import android.os.Build;
+import android.os.Bundle;
 import android.os.Environment;
 import android.preference.PreferenceManager;
 
@@ -26,6 +35,8 @@ import com.raizlabs.android.dbflow.config.FlowManager;
 
 import net.sqlcipher.database.SQLiteDatabase;
 
+import org.openmrs.mobile.receivers.*;
+import org.openmrs.mobile.receivers.ConnectivityReceiver;
 import org.openmrs.mobile.security.SecretKeyGenerator;
 import org.openmrs.mobile.utilities.ApplicationConstants;
 
@@ -42,7 +53,18 @@ public class OpenMRS extends Application {
 	private static boolean ENCRYPTED = true;
 	private OpenMRSLogger mLogger;
 
-	private Boolean isDataSyncInitialized = false;
+	// Sync constants
+	// The authority for the sync adapter's content provider
+	public static final String AUTHORITY = "org.openmrs.mobile.provider";
+	// An account type, in the form of a domain name
+	public static final String ACCOUNT_TYPE = "org.openmrs.mobile.datasync";
+	// The account name (this isn't used by the app, but is needed for syncing)
+	public static final String ACCOUNT = "defaultAccount";
+	// Sync interval constants
+	public static final long SECONDS_PER_MINUTE = 60L;
+	public static final long SYNC_INTERVAL_IN_MINUTES = 1L;
+	public static final long SYNC_INTERVAL = SYNC_INTERVAL_IN_MINUTES * SECONDS_PER_MINUTE;
+	private Account mAccount;
 
 	public static OpenMRS getInstance() {
 		return instance;
@@ -69,6 +91,53 @@ public class OpenMRS extends Application {
 
 		generateKey();
 		initializeDB();
+		registerReceivers();
+		initializeDataSync();
+	}
+
+	private void registerReceivers() {
+		// Register connectivity receiver
+		BroadcastReceiver connectivityReceiver = new ConnectivityReceiver();
+		IntentFilter connectivityFilter = new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION);
+		this.registerReceiver(connectivityReceiver, connectivityFilter);
+	}
+
+	private void initializeDataSync() {
+		// Create the dummy account
+		mAccount = createSyncAccount(this);
+		// Turn on periodic syncing
+		ContentResolver.setSyncAutomatically(mAccount, AUTHORITY, true);
+		ContentResolver.addPeriodicSync(mAccount, AUTHORITY, Bundle.EMPTY, 3600);
+	}
+
+	/**
+	 * Create a new dummy account for the sync adapter
+	 *
+	 * @param context The application context
+	 */
+	public static Account createSyncAccount(Context context) {
+		// Create the account type and default account
+		Account newAccount = new Account(ACCOUNT, ACCOUNT_TYPE);
+		// Get an instance of the Android account manager
+		AccountManager accountManager = (AccountManager) context.getSystemService(ACCOUNT_SERVICE);
+
+		// Below needed for syncing, but is for dummy account
+		if (accountManager.addAccountExplicitly(newAccount, null, null)) {
+            /*
+             * If you don't set android:syncable="true" in
+             * in your <provider> element in the manifest,
+             * then call context.setIsSyncable(account, AUTHORITY, 1)
+             * here.
+             */
+			// Intentionally left blank
+		} else {
+            /*
+             * The account exists or some other error occurred. Log this, report it,
+             * or handle it internally.
+             */
+			// Intentionally left blank
+		}
+		return newAccount;
 	}
 
 	protected void initializeDB() {
@@ -344,11 +413,7 @@ public class OpenMRS extends Application {
 		editor.commit();
 	}
 
-	public Boolean getIsDataSyncInitialized() {
-		return isDataSyncInitialized;
-	}
-
-	public void setIsDataSyncInitialized(Boolean isDataSyncInitialized) {
-		this.isDataSyncInitialized = isDataSyncInitialized;
+	public void requestDataSync() {
+		ContentResolver.requestSync(mAccount, AUTHORITY, Bundle.EMPTY);
 	}
 }
