+++ openmrs-client/src/main/java/org/openmrs/mobile/dagger/SyncComponent.java
@@ -1,5 +1,6 @@
 package org.openmrs.mobile.dagger;
 
+import org.openmrs.mobile.data.sync.PushSyncProvider;
 import org.openmrs.mobile.sync.SyncAdapter;
 import org.openmrs.mobile.data.sync.SyncService;
 
@@ -12,4 +13,5 @@ import dagger.Component;
 public interface SyncComponent {
 	void inject(SyncAdapter syncAdapter);
 	void inject(SyncService syncService);
+	void inject(PushSyncProvider pushSyncProvider);
 }
