+++ openmrs-client/src/main/java/org/openmrs/mobile/data/sync/PushSyncProvider.java
@@ -1,5 +1,6 @@
 package org.openmrs.mobile.data.sync;
 
+import org.openmrs.mobile.dagger.DaggerSyncComponent;
 import org.openmrs.mobile.data.db.DbService;
 import org.openmrs.mobile.models.Patient;
 import org.openmrs.mobile.models.SyncLog;
@@ -11,7 +12,9 @@ public class PushSyncProvider implements SyncProvider {
 	DbService<Patient> patientDbService;
 
 	@Inject
-	public PushSyncProvider() { }
+	public PushSyncProvider() {
+		DaggerSyncComponent.create().inject(this);
+	}
 
 	@Override
 	public void sync(SyncLog record) {
